requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
certifi==2025.8.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
orjson==3.11.3
prometheus_client==0.23.1
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.4.0
sniffio==1.3.1
starlette==0.48.0
structlog==25.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
